requirements.txt
asgiref==3.3.1
boto3==1.18.54
botocore==1.21.54
Django==3.1.7
django-social-share==2.2.1
django-storages==1.11.1
jmespath==0.10.0
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
asgiref==3.3.1
boto3==1.18.54
botocore==1.21.54
Django==3.1.7
django-social-share==2.2.1
django-storages==1.11.1
jmespath==0.10.0
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
asgiref==3.3.1
boto3==1.18.54
botocore==1.21.54
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==3.1.7
django-allauth==0.52.0
django-cors-headers==3.11.0
django-social-share==2.2.1
django-storages==1.11.1
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
asgiref==3.3.1
boto3==1.18.54
botocore==1.21.54
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==3.1.7
django-allauth==0.52.0
django-cors-headers==3.11.0
django-social-share==2.2.1
django-storages==1.11.1
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
wheel==0.34.2
